+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+-   Delete moderation violations from database
+-   Properly delete flagged messages if more than one were detected
+
 ## [3.0.13] - 2023-04-02
 
 ### Fixed